FIX: DataObjectInterface makes explicit the expectations of ViewableData
Subsystems relying on DataObjectInterface made use of these methods but
didn’t make it explicit.

This is essentially a bugfix on DataObjectInterface, although if someone
had implemented DataObjectInterface without ViewableData, they will
need to implement some new methods as part of this change. However,
their code is quite likely to be broken until they do so. This is also
a rarely used API.